requirements.txt
streamlit
chromadb
python-dotenv
PyPDF2
langchain
generativeai
